i need help

+++ ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: wting <wting@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/26 14:49:23 by wting             #+#    #+#             */
-/*   Updated: 2022/05/26 16:58:38 by wting            ###   ########.fr       */
+/*   Updated: 2022/05/31 10:16:04 by wting            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memset(void *b, int c, size_t len)
 {
-	int		i;
+	size_t	i;
 	char	*ntr;
 
 	ntr = (char *)b;
